rework poly_mult_div use uint8 wrapper for later

+++ pypuf/tools.py
@@ -213,11 +213,12 @@ def poly_mult_div(challenge, irreducible_polynomial, k):
     import polymath as pm
     assert_result_type(challenge)
     assert_result_type(irreducible_polynomial)
-    # challenge = challenge.astype('uint8')
-    # irreducible_polynomial = irreducible_polynomial.astype('uint8')
     # TODO Change the type to int8 or uint8
-    challenge = challenge.astype('int64')
-    irreducible_polynomial = irreducible_polynomial.astype('int64')
+    challenge = challenge.astype('uint8')
+    irreducible_polynomial = irreducible_polynomial.astype('uint8')
+    # TODO Change the type to int8 or uint8
+    # challenge = challenge.astype('int64')
+    # irreducible_polynomial = irreducible_polynomial.astype('int64')
     c_original = challenge
     res = None
     for i in range(k):
